+++ cg/services/orders/order_service/order_service.py
@@ -2,17 +2,13 @@ from cg.server.dto.orders.orders_request import OrdersRequest
 from cg.server.dto.orders.orders_response import Order as OrderResponse
 from cg.server.dto.orders.orders_response import OrdersResponse
 from cg.services.orders.order_service.models import OrderQueryParams
-from cg.services.orders.order_service.utils import (
-    create_order_response,
-    create_orders_response,
-    order_is_closed,
-)
 from cg.services.orders.order_summary_service.dto.order_summary import OrderSummary
 from cg.services.orders.order_summary_service.order_summary_service import (
     OrderSummaryService,
 )
 from cg.store.models import Order
 from cg.store.store import Store
+from cg.store.models import Order as DatabaseOrder
 
 
 class OrderService:
@@ -23,7 +19,7 @@ class OrderService:
     def get_order(self, order_id: int) -> OrderResponse:
         order: Order = self.store.get_order_by_id(order_id)
         summary: OrderSummary = self.summary_service.get_summary(order_id)
-        return create_order_response(order=order, summary=summary)
+        return self._create_order_response(order=order, summary=summary)
 
     def get_orders(self, orders_request: OrdersRequest) -> OrdersResponse:
         order_query_params: OrderQueryParams = self._get_order_query_params(orders_request)
@@ -32,17 +28,17 @@ class OrderService:
         if not order_ids:
             return OrdersResponse(orders=[], total_count=0)
         summaries: list[OrderSummary] = self.summary_service.get_summaries(order_ids)
-        return create_orders_response(orders=orders, summaries=summaries, total=total_count)
+        return self._create_orders_response(orders=orders, summaries=summaries, total=total_count)
 
     def set_open(self, order_id: int, open: bool) -> OrderResponse:
         order: Order = self.store.update_order_status(order_id=order_id, open=open)
-        return create_order_response(order)
+        return self._create_order_response(order)
 
     def update_is_open(self, order_id: int, delivered_analyses: int) -> None:
         """Update the is_open parameter of an order based on the number of delivered analyses."""
         order: Order = self.store.get_order_by_id(order_id)
         case_count: int = len(order.cases)
-        if order_is_closed(case_count=case_count, delivered_analyses=delivered_analyses):
+        if self._order_is_closed(case_count=case_count, delivered_analyses=delivered_analyses):
             self.set_open(order_id=order_id, open=False)
         elif not order.is_open:
             self.set_open(order_id=order_id, open=True)
@@ -52,8 +48,39 @@ class OrderService:
         return OrderQueryParams(
             page=orders_request.page,
             page_size=orders_request.page_size,
-            offset=orders_request.offset,
+            search=orders_request.search,
             open=orders_request.open,
             customer_id=orders_request.customer_id,
             workflows=[orders_request.workflow],
         )
+
+    @staticmethod
+    def _create_order_response(order: DatabaseOrder, summary: OrderSummary | None = None) -> Order:
+        return Order(
+            customer_id=order.customer.internal_id,
+            ticket_id=order.ticket_id,
+            order_date=str(order.order_date.date()),
+            id=order.id,
+            is_open=order.is_open,
+            workflow=order.workflow,
+            summary=summary,
+        )
+
+    def _create_orders_response(
+        self, orders: list[DatabaseOrder], summaries: list[OrderSummary], total: int
+    ) -> OrdersResponse:
+        orders: list[Order] = [self.create_order_response(order) for order in orders]
+        self._add_summaries(orders=orders, summaries=summaries)
+        return OrdersResponse(orders=orders, total_count=total)
+
+    @staticmethod
+    def _add_summaries(orders: list[Order], summaries: list[OrderSummary]) -> list[Order]:
+        order_map = {order.id: order for order in orders}
+        for summary in summaries:
+            order = order_map[summary.order_id]
+            order.summary = summary
+        return orders
+
+    @staticmethod
+    def _order_is_closed(case_count: int, delivered_analyses: int) -> bool:
+        return delivered_analyses >= case_count
